separate crawl logic
